music_render: use 32-bit to save 98 bytes
I moved as much of the low-hanging fruit to 32bit registers as possible.
This code was really written for n32 and it's nice 64-bit registers, but
we are stuck on o32.

I also made an effort to minimize stack usage by moving param loads
closer to their usage.

When hitting cache, loads are basically as fast as registers, with only
one delay slot.